pyproject.toml
[tool.poetry]
name = "redis-sansio"
version = "0.1.0"
description = ""
authors = ["Sean Stewart <sean_stewart@me.com>"]

[tool.poetry.dependencies]
python = "^3.7"
async-timeout = "^4.0.2"
aioredis = "1.3.1"
uvloop = "^0.16.0"
attrs = "^21.4.0"

[tool.poetry.dev-dependencies]
yappi = "^1.3.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
